Fix balance score calculation to handle decimal timer values
- Changed calculate_balance_score_ajax to parse float values instead of int
- Updated calculate_single_leg_balance_score type hints to accept float parameters
- Fixes 400 Bad Request error when balance timer records decimal seconds

+++ apps/assessments/scoring.py
@@ -347,7 +347,7 @@ def calculate_pushup_score(gender: str, age: int, reps: int, push_up_type: str =
         return _fallback_pushup_score(db_gender, age, reps, push_up_type)
 
 
-def calculate_single_leg_balance_score(right_open: int, left_open: int, right_closed: int, left_closed: int) -> float:
+def calculate_single_leg_balance_score(right_open: float, left_open: float, right_closed: float, left_closed: float) -> float:
     """
     Calculate single leg balance score based on time in seconds.
     Uses database standards with fallback to hardcoded values.
